+++ webpack/postcss.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   plugins: {
-    'postcss-cssnext': {
+    'postcss-preset-env': {
       browsers: 'last 3 versions'
-    },
+    }
   },
 };